requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
chromadb==0.4.15
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
requests==2.31.0
google-api-python-client==2.86.0